+++ awc/CHANGES.md
@@ -3,10 +3,13 @@
 ## Unreleased - 2021-xx-xx
 - Rename `Connector::{ssl => openssl}`. [#2503]
 - Improve `Client` instantiation efficiency when using `openssl` by only building connectors once. [#2503]
-- `ClientRequest::send_body` now takes an `impl MessageBody`. [#????]
+- `ClientRequest::send_body` now takes an `impl MessageBody`. [#2546]
+- `impl Future` for `JsonBody` no longer requires the body type be `Unpin`. [#2546]
+- `impl Stream` for `ClientResponse` no longer requires the body type be `Unpin`. [#2546]
+- `ClientResponse` is no longer `Unpin`. [#2546]
 
 [#2503]: https://github.com/actix/actix-web/pull/2503
-[#????]: https://github.com/actix/actix-web/pull/????
+[#2546]: https://github.com/actix/actix-web/pull/2546
 
 
 ## 3.0.0-beta.14 - 2021-12-17